test(vitest): port chatRuntimeSessionContext to Vitest
